requirements.txt
APScheduler==3.8.1
cachetools==5.0.0
certifi==2021.10.8
imutils==0.5.4
numpy~=1.22.1
opencv-contrib-python==4.4.0.46
python-telegram-bot==13.10
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tornado==6.1
tzdata==2021.5
tzlocal==4.1
cv~=1.0.0
opencv-python~=4.5.5.62
python-dotenv~=0.19.2